CHanged Non-Active-Time model to save dates as integers in the DB.

+++ src/modules/graphql/resolvers/non-active-time.resolver.ts
@@ -1,4 +1,4 @@
-import { Resolver, Query, Mutation } from '@nestjs/graphql';
+import { Mutation, Query, Resolver } from '@nestjs/graphql';
 import { Asset, checkAndGetBasePermission, DBOperation } from '../../permissions/permission.interface';
 import { Get } from '../../../utils/get';
 import { NonActiveTimePersistenceService } from '../../persistence/non-active-time.persistence.service';
@@ -9,25 +9,25 @@ export class NonActiveTimeResolver {
 
   @Query('nonActiveTimes')
   async getNonActiveTimes(_, {}, context) {
-    // checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.READ, Asset.ANY);
+    checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.READ, Asset.ANY);
     return this.nonActiveTimePersistence.getAll();
   }
 
   @Mutation('createNonActiveTime')
   async createNonActiveTime(_, { nonActiveTime }, context) {
-    // checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.CREATE, Asset.ANY);
+    checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.CREATE, Asset.ANY);
     return this.nonActiveTimePersistence.createNonActiveTime(nonActiveTime);
   }
 
   @Mutation('updateNonActiveTime')
   async updateNonActiveTime(_, { id, nonActiveTime }, context) {
-    // checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.UPDATE, Asset.ANY);
+    checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.UPDATE, Asset.ANY);
     return this.nonActiveTimePersistence.updateNonActiveTime(id, nonActiveTime);
   }
 
   @Mutation('deleteNonActiveTime')
   async deleteNonActiveTime(_, { id }, context) {
-    // checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.DELETE, Asset.ANY);
+    checkAndGetBasePermission(Get.getObject(context, 'user'), DBOperation.DELETE, Asset.ANY);
     return this.nonActiveTimePersistence.deleteNonActiveTime(id);
   }
 }
